requirements.txt
google-cloud-bigquery
apache-beam[gcp]
apache-airflow[gcp]
psycopg2-binary
pandas
sqlalchemy
apache-airflow
loguru
jupyter
